Change styling of SectionCards.tsx
Added some more properties to the Section Cards template, including
upper border on each child to work as a separator, and also bg colour
to differentiate between them. Ultimately I think it would be ideal
if we can incorporate team colors to the background for each child,
I have manually set FSU color in this instance to see how it looks.