Compile together TeamLandingPageComponents file with all elements
- Previously built out the components directly on the landing page, just
to initially get bearings.
- I have now built these components out on the TeamLandingPageComponents
file and importing them into the TeamLandingPage.tsx file, this makes it
far more readable.
- Stats component has not yet been configured, and also the TeamOverview
component is not complete either, I have to implement logic to not
reveal CFB players overall numeric value when switching to SimNFL, and
also need to add data for remainder of the component.