Further work on CFB Team page
- Created a function called isBrightColor which checks if the teams
colorOne is bright (thresholds set) if it is then it swaps it for the
secondary color, this is to keep a better contrast on roster page and,
ive also implemented this elsewhere i.e. on the landing page.
- Tailwind classes added for screen breakpoints namely apple, samsung.
- Table developed further to correctly use selectedTeam colors.
- Team Info component adapted to better stack and incorporate <Text>
- Logo now enclosed in a border and background related to team colors,
also utilising darkenColor for the slim border around logo.
- CFB Team Page mobile responsiveness optimised. Certain attributes have
been omitted to give a better presentation on mobile and fit within the
constraints of the screensize.